+++ tools/fs.h
@@ -29,14 +29,14 @@ namespace evo{
 				EVO_NODISCARD inline auto getHandle() const noexcept -> ::FILE* { return this->handle; };
 
 
-				EVO_NODISCARD auto readLine() const noexcept -> std::optional<std::string>;
+				EVO_NODISCARD auto readLine() const noexcept -> evo::Result<std::string>;
 
 				EVO_NODISCARD auto writeLine(const std::string& text) noexcept -> bool;
 
 
 
-				EVO_NODISCARD auto read() const noexcept -> std::optional<std::string>; // reads all
-				EVO_NODISCARD auto read(size_t data_size) const noexcept -> std::optional<std::string>;
+				EVO_NODISCARD auto read() const noexcept -> evo::Result<std::string>; // reads all
+				EVO_NODISCARD auto read(size_t data_size) const noexcept -> evo::Result<std::string>;
 
 
 				EVO_NODISCARD auto write(const std::string& data) noexcept -> bool;
@@ -59,8 +59,8 @@ namespace evo{
 			public:
 				EVO_NODISCARD auto open(const std::string& path, Flags<FileMode> mode) noexcept -> bool;
 
-				EVO_NODISCARD auto read() const noexcept -> std::optional< std::vector<byte> >; // reads all
-				EVO_NODISCARD auto read(size_t data_size) const noexcept -> std::optional< std::vector<byte> >;
+				EVO_NODISCARD auto read() const noexcept -> evo::Result< std::vector<byte> >; // reads all
+				EVO_NODISCARD auto read(size_t data_size) const noexcept -> evo::Result< std::vector<byte> >;
 
 				EVO_NODISCARD auto write(const std::vector<byte>& data) noexcept -> bool;