+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Evo Changelog
 
+### v1.4.0
+- Changed methods in `fs::File` and `fs::BinaryFile` that returned `std::optional` to now return `evo::Result`
+- Changed `evo::stringsEqual()` to now return `evo::Result<bool>`
+- Added to and reorganized README
+
 ### v1.3.1
 - Fixed documentation links